SWITCH TO SERVER-UPLOAD. Use FileReader, native JS
NOTE: After this commit, all changes on this branch will be trying to
upload the images server-side (so the browser uploads to the server
which then uploads to AWS). This is because uploading on the
client-side using pre-signed URLs is producing too many problems.
Use the native FileReader object from JavaScript to convert files to
text strings.
WARNING: This commit still cannot handle uploads without an error. AWS
is responding with a 403 Forbidden, saying that the signature provided
doesn’t match the one they calculated.